feat: Add pack support to CStruct
This change introduces the ability to specify the 'pack' attribute for CStruct classes, which controls the memory alignment of the structure's fields.

This is achieved by passing a 'pack' argument to the CStruct class definition. The implementation forwards the 'pack' value to the underlying ctypes.Structure.

Unit tests have been added to verify that the 'pack' attribute correctly affects the size of the structure.